move pages folder to /src
